app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
gunicorn>=21.2
psycopg2-binary>=2.9
black>=24.3.0
pytest>=8.0.0
pytest-django>=4.8.0
requests>=2.32.0
